display the transactions in a block
